Undef serialize for intrinsics conflict
Summary:
In g++, there is some namespace that is defining macros that conflict with the definitions included within this file. Andrew suggested this as a potential workaround -- I don't see this as the most maintainable fix to get this on g++, but it unblocks the barnfs workstream. I also think it is probably safe, because the scope of the undef is within the same file.

The alternative solution is to figure out how the _serialize macro is leaking in here, but I don't know how to do that. If anyone has suggestions, I'm happy to look into this.

Reviewed By: ahilger

Differential Revision: D64269897

fbshipit-source-id: 902217fe33e6db9cff86cac9778db160f99bee32