remove old depgraph reader code
Summary:
the old depgraph format no longer fits in a 4GB file limit, so the
32-bit indexes do not work. Henri disabled the codepath some months ago, and
deleted the flag to write the old depgraph format.

This diff just deletes the code to read the old format. I moved enums into
tuple structs, and just deref to get the fields inside. I'm sure there's some
extra cleanup to delete the traits on the inner fields, and maybe flatten the
whole thing but for now this just removes dead code.

Reviewed By: mjhostet

Differential Revision: D64401433

fbshipit-source-id: d626a1513c470a77da89eb08bf35c28e268fcc11