Change legacy reflection to return non-constexpr type name
Summary:
The type name is constexpr in legacy reflection.

However, in always-on reflection we changed to non-constexpr `std::string_view` to save build time, we need to change the legacy reflection API to non-constexpr as well. This diff changed the type name to return non-constexpr std::string instead.

The reason we don't return std::string_view is that some functions require concatenation which is hard to do with std::string_view

```
  static std::string rname() {
    return "map<" + key_helper::name() + ", " + mapped_helper::name() + ">";
  }
```

Reviewed By: rmakheja

Differential Revision: D64342107

fbshipit-source-id: 548854c4d9744f89dc13ca7e81366c8d617338a9