Separate read only environment
Summary:
As a result of threading environment through we were forgetting to reset the `for_` flags in the environment. Consequently, we were overconstraining program generation and restricting variety.

This is a classic mutation problem. To remediate it, this diff separates the readonly part of the environment from the mutable part that gets threaded through. So we can only shadow the readonly values while recursive but not carry them over to adjacent calls through returns.

Reviewed By: mheiber

Differential Revision: D64471606

fbshipit-source-id: 9214d0f88d27f6a26b987135d3f656dd6e27b836