Fix crash when TilePromise is fulfilled after connection is destroyed
Summary: There is a heap-use-after-free when the `fn` in `HandlerCallbackBase::fulfillTilePromise` runs because it is possible that the connection has already been destroyed by the time the EventBase executes `fn`. This is because the destruction of the connection is guarded by "inflight requests" which gets decremented after the response is sent to the client from the EventBase. The response gets added to the EventBase through a custom ReplyQueue which doesn't limit max number of items processed in the event loop but the tile fulfillment gets added to the EventBase using the general queue which does limit max number of items processed in the event loop. As a result, the response task gets completed before the tile fulfillment task and allows the connection to be destroyed before the tile is fulfilled. To fix, we can enqueue the tile fulfillment using the same custom queue to ensure that the tile is always fulfilled before the response is sent.

Reviewed By: yfeldblum

Differential Revision: D64521715

fbshipit-source-id: 8dceedae53a9a88f409931601fd2e093dc74b521