Fix jmpi implementation
Summary:
It was trying to be smart and emit a direct jump if it could fit in the
PC-relative offset.  However, even though that may succeed at emission time, the
code will be relocated later and may fail because the offset now is to large.
This diff removes that handling and just emits a far jump sequence, which may be
optimized later during relocation.

Reviewed By: ricklavoie

Differential Revision: D64612558

fbshipit-source-id: 04a770261a305d0097e440dc850799740c6788c0