Get rid of duplicate functionality between expr_of and is_immediately_inhabited
Summary:
`is_immediately_inhabited` was almost exactly the same as `expr_of` if we squint hard enough. The former was `Option.is_some` applied to the former.

This diff gets rid of that duplication.

Caveat: `is_immediately_inhabited` is deterministic, but `expr_of` is not. I changed `expr_of` so that whether it returns `Some` or `None` is deterministic, but not what expression i returns. This works because the predicate usage of `expr_of` inside `subtype_of` does not care about the actual expression just that there would be one.

Reviewed By: mheiber

Differential Revision: D64572907

fbshipit-source-id: e9fc47274f534a1b9faaa6c872ba75a5e9f7a8f9