Store derivation in user error
Summary: Stores an additional explanation of a subtyping error, generated from extended reasons, in the user error datatype. This allows us to generate the message without restarting hh. The new field is invisible to equality and comparison so we will still only get one error if the explanations differ  - this only happens due to a quirk with bi-directional typing / expected types in function call typing

Reviewed By: madgen

Differential Revision: D64401333

fbshipit-source-id: 92988cf219b56fe8eb14a7ccb10df2e9ea96f0fa