Add new HandlerCallback type for oneway
Summary:
The relevant changes are all in `thrift/lib/cpp2/` (most notably `AsyncProcessor.h`)

Today, `oneway` uses `HandlerCallbackBase` which has `exception` but not `done` which means that `ServiceInterceptor::onResponse` is executed only on exceptions but not on successful completion of a oneway request.

This is surprising so this change essentially adds a `done` method for the callback passed to oneway methods.

This means that for oneway methods, we now pass in a subclass of `HandlerCallbackBase` named `HandleCallbackOneWay`. This aligns with non-oneway methods where we pass in `HandlerCallback<T>`. The end result is that `HandlerCallbackBase` is not explicitly constructed anywhere (as its name suggests it shouldn't).

The API of `HandlerCallbackOneWay` is similar to `HandlerCallback<void>`. The could potentially be merged together but requires a larger refactory.

Reviewed By: yfeldblum

Differential Revision: D64195501

fbshipit-source-id: fefaa2f201ed1b98ae1ea46c1ae486817a29cb9e