Mark folly_sdt_semaphore as needing name based resolution
Summary:
folly_sdt_semaphore is referred to by name in inline assembly.  This particular
inline assembly use cannot be converted to a constraint template parameter
(until llvm 18 / gcc 14), as their is no way (until those versions) to specify
that a non-pic relocation is needed when compiling under pic.  This inline
assembly requires a non-pic relocation because it is being written to the
`.notes` section, which is non .text.

Reviewed By: jpporto

Differential Revision: D64604573

fbshipit-source-id: c328b20e5e85aa0896a59d563443b2cede49cc59