Temporarily implement enums along with abstract types.
Summary: Until we generate a separate file for enums, generate enums in the thrift_abstract_types module as well. For now this prevents some test failures. This will probably start to fail when tests that exercise inheritance start to come into play. At that point, create the new enum file.

Reviewed By: yoney

Differential Revision: D64145572

fbshipit-source-id: ea89db252f9ce14e656dd072cabb9f7486d0c444