defining custom compressor
Summary:
* Motivation for adding custom compressor support - https://fburl.com/gdoc/h2kwnv6u
* Overview of the rest of the plan - https://fburl.com/gdoc/7tkibmzh

Before we can start with the implementation, we need to make custom compressor play nicely with open source. This means support a special 'CUSTOM' compressor, and that users would be able to register their own custom compressors.

The first problem is that: unlike `ZLIB`, `CUSTOM` by itself is not sufficient to narrow down *which* custom compressor it is using.

In this diff, we simply add a corresponding `optional string customCompressorId` whereever we currently have an enum `CompressionAlgorithm`. There needs to be corresponding data plumbing to make sure we can actually have access to the new `customCompressorId` everywhere, but that should be easier to do.

An alternative solution is presented by cevans87 in D63356848 stack. The basic idea is to achieve the same purpose by using `CodecConfig` instead. The pros and cons are as following:

* Simply adding `customCompressorId` - this diff
  * [+] simple and more backward compatible - we don't need to migrate away from the current `compression` enum.
* Adding a new `CodecConfig` - D63356848
  * [+] encapsulates related fields into the same struct, which is just a good design pattern and make it more future proof
  * [+] even though no existing use case needs it, it is very general in the sense that it allows compression/decompression to take additional arbitrary configurations. That seems very future proof. For existing use cases, we just need a simplified `CodecConfig` where we just need to identify the name of the custom compressor.
  * [-] we have concerns whether it would increase the payload size as it needs to be send per request
  * [-] since we would need to maintain backward-compatibility, it might make the codebase more complex

### Update 10-11
Based on discussion with robertroeser, we simplified the custom compression type to simply be an enum (`CUSTOM`) without the compressor id. The assumption is that:
* within Meta, we know what this 'custom' compressor is going to be used for
* if the use cases arises (i.e. if we need to register more than one custom compressor), we can additionally add the compressor id without undoing any of the changes here.
* this makes it much easier for us to develop a proof-of-concept integration without having to change a lot of codes

Reviewed By: robertroeser

Differential Revision: D63768836

fbshipit-source-id: 854d7c30a4c6c41c6d3ac5260169a2fa66fa033e