Ban cyclic bound on type aliases
Summary:
This diff is about cyclic bounds on type aliases like:
```
netype A as X<A> = ...
```
These are actually already "banned", though not explicitely. By that I mean that the bound will be localized to `X<mixed>` (mixed arises from detecting a cycle) and typing will subsequently fail in a confusing way. Instead, this diff raises a clear error stating that this cycle bound is not allowed.

We only do that for `as` constraints for now, and don't change anything for `super` constraints. `super` constraints are treated similarly to the RHS of case types, and we support cyclic definition there, so there's not reason we cannot support cyclic `super` constraints as well.

Differential Revision: D64479259

fbshipit-source-id: 7741ad8208e59771f95be7f4fc9f55adcfbd31ca