Typing_case_types.get_variant_tys - return None for case types with where clauses
Summary:
This function effectively expands a case type to its variant hints (generally to be used as a union) and is used in two places:
* https://fburl.com/code/vg20lcs4 (Typing_intersection)
* https://fburl.com/code/r4m41zte

In the Typing_intersection site, it's used to see if `T & CaseType` can be simplified based on the definition of `CaseType`. For case types with `where` clauses, it's not correct to ignore the `where` clauses here since they may not be satisfied. For this use case we'd need to evaluate the where clauses to pick out the valid variant hints. I'm not sure how feasible this is. But for now, I believe it's sufficient to just skip the where-clause case (not try to simplify the intersection in the intersection use case) and I moved/updated the TODO to mention this avenue of exploration.

I noticed this issue when writing code that tried to split (type switch) such a case type. The splitting logic creates intersections with the case type, but we were losing them in the simplifications. This meant that we lost the fact that the value was a value of the case type in both branches of the type switch. For regular case types, this is fine, but for case types with where clauses, the constituent variant hints aren't unconditionally subtypes of the case type so it's important to keep it around (see added test case)

Reviewed By: CatherineGasnier

Differential Revision: D64497813

fbshipit-source-id: d8b6cf544955806c7c9d8a708ad1779ec5eb0358