Allow ServiceInterceptor::onConnection to throw and terminate connection
Summary:
This diff allows `ServiceInterceptor::onConnection` to throw exceptions in order to abort the connection.

## Why might this be useful?

For ACL checking, we can sometimes detect that a connection is "bad" and will always fail ACL checks. Rather than having every request fail, it would be better to close the connection ASAP so that client libraries can retry or report an error where most appropriate.

## Why wasn't this there in the first place?

Primarily because (as seen in this diff), only Rocket transport really has the capability to report errors on connection failure. But it seems like it's a very helpful capability and also Rocket is (almost) everywhere now.

Reviewed By: robertroeser

Differential Revision: D64719072

fbshipit-source-id: 56689c2878300913b24406e79c4ab78488624dc0