Remove special purpose parse error for $x{e}
Summary:
Remove parser support for building a ffp for Subscript with PHP style `{` and`}` braces. This means we can't give a targeted error message anymore, but often in the presence of other parse errors, the ffp would create a Subscript expression out of what should have been statement-level braces, leading to confusing error messages.

We generally don't show the errors from rust_parser_error, when there were errors in the ffp, so the above wasn't a big problem, but we'd like to start showing them since they are often more understandable.

Reviewed By: vassilmladenov

Differential Revision: D64757894

fbshipit-source-id: b5e606c11cf5331c6dd632864fb5bfa2a2d1def0