Create an opportunity for early termination
Summary:
`subtype_of` has a `step` that can recurse (potentially very deep) or succeed quickly. Later on among many choices we pick a random one which might simply discard the deeply recursive result!

To prevent this waste, We introduce a choice point between non-recursive and recursive choices in `step` if the former is chosen, we skip a lot of computation.

Reviewed By: CatherineGasnier

Differential Revision: D64861946

fbshipit-source-id: b9b431780c1cdbdaf8dc1d331ac796c9afa5fd29