Clarify error message of linter about filename mismatch
Summary: We have a linter checking that a file name is consistent with its class name if the file contains only one class. If the file also contains type definitions, these are ignored by the linter. However in that case the error message "file.php only contains MyClass, it should be named MyClass.php" is confusing, because the file does not "only contain MyClass". This diff adapts the error message to make it less confusing.

Reviewed By: madgen

Differential Revision: D64845159

fbshipit-source-id: b51a5828428210cee5f13600360558f4b3a7319e