Enable TC allocation compatible with ROAR
Summary:
This introduces a weak symbol that controls the start of the HHVM translation
cache.  By default this uses `sbrk(0)` (matching the current behavior).

Under ROAR `sbrk(0)` does not return the correct value, because ROAR adds
metadata to the binary that lives in high memory.  This makes sbrk return very
high addresses which are unsuitable for the TC.  ROAR leaves plenty of space in
low memory so just read tc_start_address reads the symbol roar uses to mark the
start of this free space.

While here configure the size of the ROAR TC to be 500MB.

Reviewed By: ricklavoie, ottoni, jpporto

Differential Revision: D64799159

fbshipit-source-id: 097995f0dc318c365186b1f7e62242a824662bda