Use whisker for Go
Summary:
Whisker is a replacement for mustache and is largely backward compatible. It brings several quality of life improvements like... error checking.

You can read more about the details [in this post](https://fb.workplace.com/groups/498666380550247/permalink/2051389461944590/).

This diff updates the Go codegen to use Whisker instead of mstch. The fixtures are mostly unchanged, except for bug fixes in Whisker related to whitespace handling (you can see it in the fixtures).

Reviewed By: Filip-F

Differential Revision: D64860317

fbshipit-source-id: 3b8c9d81d14eb9f7386e63291736b65bbd898785