Remove __call__ from thrift abstract types.
Summary: `__call__` must be interface compatible with the mutable API `__call__` that, for container fields, must accept either a Mutable container or the result of a `to_thrift_*()`. This will not match the immutable interface, which accepts `container.abc` types. Therefore, there is no common interface to abstract. Therefore, at this time, abstract must not provide `__call__`.

Reviewed By: yoney

Differential Revision: D64479137

fbshipit-source-id: cb6f743f6694b1ebb52cdb110e2c6d3118772e45