distinguish user from internal errors
Summary:
## What

For converting hack files to notebooks, when something goes wrong we want to distinguish:
- user's fault, such as bad metadata or ill-formed Hack: give zero exit code and an error message
- our fault, in which we give a nonzero exit code and an error message

## Why

Since Hack files can be edited by humans, codemods, etc. we want to be clear about when a problem is the user's fault vs. ours–so the user knows to report the problem if it's on us or fix it if it is on them.

Error codes also show up in our telemetry, so distinguishing these cases could help guide how to improve.

## Context

for more information, see fb-only section below

Reviewed By: patriciamckenzie

Differential Revision: D64826947

fbshipit-source-id: 25f88f9f7d8c88ddf636e9a224c19c0624ebce7d