Look for specific patterns in hhstc output
Summary:
So far we were only interested in generating well-typed programs and hence looking for "No errors" output from hhstc.

This diff parametarises that pattern out of the script. The goal is to be able to add other tests that introduce errors predictably and check that the appropriate error output is generated in each case.

Reviewed By: mheiber

Differential Revision: D64900595

fbshipit-source-id: 297eb21d9fd1eda7b53610fdb6d1c52d40fa4acd