Fix --type-at-pos-batch bug w/ mixed splice/non-splice ET positions
Summary: `--type-at-pos-batch` asks for types at multiple positions. If there is a type in a splice and type not in a splice but both in the same expression tree, we want to report the non-virtualized type for the former and the virtualized for the latter. Previously, we were incorrectly handling it assuming all or nothing.

Reviewed By: CatherineGasnier

Differential Revision: D64939267

fbshipit-source-id: b079e1626d701b632e5bd4c0141abd5fa8652af9