lower naively interceptjcc to jcc
Summary:
When `fast_fbintercept2` is enabled, `interceptjcc` is used to check surprise flags in the lowering of `CheckSurpriseFlagsEnter`.  The `fast_fbintercept2` binary rewriting logic expects its condition to always be `CC_AE` and successful target equal to the `handleSuprise` label.  It also assumes that it is assembled naively to `jae`.

The `vasm-x64` `interceptjcc` emitter however relied on the generic `jcc` translation, which in some cases switches condition and labels to optimise away a jump.  This diff ensures that  `interceptjcc`  is emitted "naively", without performing the `jcc` optimisation that can invalidate the `fast_fbintercept2` logic.

Reviewed By: ottoni

Differential Revision: D62300184

fbshipit-source-id: 646303a2dcf84179b36108b87b6d646d74be83eb