Remove unused-but-set variables in thrift/compiler/generate/t_mstch_java_generator.cc +1
Summary:
Required for LLVM-19 upgrade.

This diff removes a variable that was set, but which was not used.

LLVM-19 has a warning `-Wunused-but-set-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused but set variables often indicate a programming mistake, but can also just be unnecessary cruft that harms readability and performance.

Removing this variable will not change how your code works, but the unused variable may indicate your code isn't working the way you thought it was. If you feel the diff needs changes before landing, **please commandeer** and make appropriate changes: there are hundreds of these and responding to them individually is challenging.

For questions/comments, contact r-barnes.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: palmje

Differential Revision: D64996202

fbshipit-source-id: 964b99a1a06cce2ac4e959e6a56daeb0d2f2d857