Log bytestream when connection is thought to not be thrift.
Summary:
Check for all HTTP method signatures, roll them into a single return value of "HTTPAttempt".  The code also now prints the first 48 bytes as a string when an HTTPAttempt is found (in a previous patch).

Document all of the ClientType values with comments.

When both the first 32 bit word and the second 32 bit word of the connection are unknown to thrift log 48 sanitized bytes along with all 48 bytes in hexadecimal in the log to help indicate what is going on.

Reviewed By: awalterschulze

Differential Revision: D64850461

fbshipit-source-id: 3b8dccb0b72400c5f07328d288605beee3ac32b2