`coro/TaskWrapper.h`, a helper for wrapping `Task` / `TaskPromise`
Summary:
Here are few possible reasons to want to wrap `Task`:
 - `SafeTask` (upcoming in D61995850) -- quacks exactly like `Task` but includes some additional compile-time enforcement that makes it harder to make dangling-reference bugs.
 - `TaskDfatalIfNotAwaited` -- quacks exactly like `Task`, but used for those cases where you really want to make sure the task is awaited.

It would be possible to open up `Task` / `TaskPromise` for inheritance, but this comes with a risk of object slicing bugs, some quite subtle. Moreover, while for the simplest wrappers the "is-a-Task" implied by relationship can be okay, one can imagine more featureful wrappers where it is absolutely not appropriate. The composition approach is more explicit, and thus safer for the implementer of the wrapper.

The downside of composition is that we have to transparently forward the entire -- and rather large -- API surface of `TaskPromise` / `Task`. This would be far too fragile if it were copy-pasted in each individual wrapper. By collecting the implementation in a central `TaskWapper.h`, the cost of adapting to `Task.h` API changes should stay low.

Reviewed By: ispeters, skrueger

Differential Revision: D62994547

fbshipit-source-id: 5d7e11a386950b9e07d34b28cf7f986f7eb94e66