Implement shape destructure in hackc
Summary:
Allow a shape as an lvalue, in analogy with a list.

This is implemented in the emitter, where we generalise the notion of index into the list from int to a union of int and shape_field_names. Then we can either generate an integer member key (EI), or a string member key (ET) in the simple cases. In more complicated cases such as class constants and nameof, we might need to generate some code to load the key onto the stack, and then we use an (EC) member key.

Reviewed By: CatherineGasnier

Differential Revision: D65145905

fbshipit-source-id: 9f033223bc6096fdf8162ba633855ed907d74696