Remove propagation of profiled dict keys from IterInit to IterNext
Summary:
We are no longer trying to have a shared iterator header between IterInit and
IterNext, so their guards don't need to match.

If we guarded dict keys at IterInit, there is no point of repeating this check
at IterNext. Either the check at IterInit is dominating during the whole loop
and thus redundant at IterNext, or it is not and it is better to guard the
actual observed keys.

Note that currently this CheckDictKeys optimization is ineffective and does not
propagate to key usage. This will be fixed later.

Reviewed By: ricklavoie

Differential Revision: D65225120

fbshipit-source-id: 2cdf45b517e6f54aa0c1dc2fcfdafc320840794f