Separate IterGetKey and IterGetValue out of IterInit and IterNext
Summary:
Remove the key and value extraction from IterInit and IterNext, move it
to the new IterGetKey and IterGetValue bytecodes.

The main motivation is to undo the complexity and fragility of the logic that
attempts to share IR between IterInit and IterNext. This diff reduces the
amount of shared logic to storing the current iterator position. The next diff
eliminates it altogether.

Mostly a straightforward refactor with the following notes:

- for objects implementing the Iterator interface, we just call key() and
  current() directly from the JIT, since it was easier than adding a new IR
  instruction (and the code runs faster too!)
- the CheckDictKeys optimization became temporarily no longer effective,
  losing information about guarded key types; however, hhbbc is now able to
  insert AssertRATStk before we set the key/value locals, which together with
  improved sinking of IncRefs seems to avoid a perf regession; CheckDictKeys
  optimization will be addressed in future diffs (likely via type system
  improvement)
- hhbbc BaseConst optimization required an extra tracking whether the top stack
  element is equivalent to the current iterator key

Reviewed By: ricklavoie

Differential Revision: D64852480

fbshipit-source-id: ab8f9196e2defe2548209f0d84425a5ebeaedceb