Increase Parallelism of ZSTD_DCtx Pool to 16
Summary:
An internal use case was reporting significant contention on the lock in the
pool that backs the "core-local" cache for `ZSTD_DCtx`es. Evidently, 4 cache
slots is not enough:

https://fb.workplace.com/groups/881214461949572/posts/27364067066570951/

The tradeoff here is that increasing the number of slots means that an
application may create more contexts that will remain resident in memory even
if they're not being used.

We had to roll back a previous change (D29944813) that tried to increase this
count and ran into a memory usage explosion. But that diff also changed the
`ZSTD_CCtx` pool size, and `CCtx`es can get very large indeed, while `DCtx`es
are much more well behaved.

Reviewed By: terrelln

Differential Revision: D64837870

fbshipit-source-id: c7851170b6539063a90d1b8cc53ffab0e70e9163