Add interaction terminal errors wiki page
Summary:
There are some pre-existing SR error reasons which effectively mean the interaction is no longer valid. There is also a new `INTERACTION_LOADSHEDDED` error reason which is used to mark an interaction as terminated if an interaction request was load-shedded.

Clients should handle these errors by recreated but this is not documented in our wikis. Lets add documentation.

Reviewed By: islamismailov

Differential Revision: D65231744

fbshipit-source-id: 74bee881bbb3b2afd51ece38d226b3f44f6b4eaf