Ensure uniform python class names for the FbThriftUnionFieldEnum
Summary:
Given an union named `Foo`, he python class name for the `FbThriftUnionFieldEnum` is
Immutable API: `Foo`
Mutable API: `FbThriftUnionFieldEnum_Foo`

With the introduction of a common Abstract API, the names must be consistent. Given the prevalent use of the immutable API, change the name in the mutable API to match the immutable one.

Reviewed By: yoney

Differential Revision: D65341693

fbshipit-source-id: ff0c0287338401f1222844ac4e890f05e8e5081d