Proper canonicalization during package config processing
Summary:
when parsing the package config, we
1. replace "//" at the start of include_paths with the parent directory of PACKAGES.toml
2. strip the root from the path we get in step 1

Both the root and the package config parent directory used to be canonicalized during this process. D63693534 added a `strict` mode and removed the canicalization step from the parent directory but kept it for root; as a result step 2 above would not yield the correct include_paths for `get_package_for_file` matching.

This diff makes it so that we canonicalize **both** root and parent directory when `strict` is true, and **neither** when it's set to false.

Reviewed By: DavidSnider

Differential Revision: D65355822

fbshipit-source-id: 409e1578a0a69798c44cfae9ab57377945c2c924