Add gating classes to help switch from MySQL protocol to Thrift protocol (take 2)
Summary:
This is a second attempt at this diff and combines the original diffs of D57460644, D63141477 and D63267393.

Original text:

In order to make client migration from MySQL protocol to Thrift protocol as seemless as possible, we are making the top level FbAsyncMysqlClient, FbAsyncConnectionPool, FbSyncMysqlClient and FbSyncConnectionPool into gating classes.  This means renaming the existing clients and replacing them with the gating clients (the gating clients only choose between MySQL protocol and Thrift protocol).
The following get renamed:
- FbAsyncMysqlClient -> FbAsyncMsyqlProtocolClient
- FbAsyncConnectionPool -> FbAsyncMysqlProtocolConnectionPool
- FbSyncMysqlClient -> FbSyncMysqlProtocolClient
- FbSyncConnectionPool -> FbSyncMysqlProtocolConnectionPool
Next the original clients are recreated with code that checks the gating status and either routes request to the renamed class or to the FbThriftMysqlClient class.
Note that Thrift doesn't have expensive connections so there is no need to have connection pooling, so the connection pool versions go to FbThriftMysqlClient - not to some pooled version of the thrift client.

In the original version the FbThriftMysqlClient class included a FunctionScheduler, which is a fairly expensive object.  The purpose was to provide for a background function that would run occasionally.  Unfortunately, when lots of FbThriftMysqlClient objects get created this greatly increased the memory needed for the binary.  The FunctionScheduler has been removed and instead a cancellable coroutine runs in a loop and is cleaned up on destruction.  This is much less expensive.

In addition, when creating both a client and a pool from the client, we were mistakenly creating two FbThriftMysqlClient objects rather than reusing the one from the client in the pool.

Reviewed By: fadimounir

Differential Revision: D63903128

fbshipit-source-id: 04df32917408ab4dd15528770f784cf461e987c4