Unify all type constraints into TypeIntersectionConstraint
Summary:
Refactor TypeIntersectionConstraint to contain both the type hint constraint as well as the upperbounds.

Currently, the constraint coming from the type hint is stored as typeConstraint or retTypeConstraint for return types. The "upperbounds" are stored in a separate struct called TypeIntersectionConstraint. But this distinction is unnecessary and makes code complicated. Unify all the type constraints into TypeIntersectionConstraint.

Reviewed By: jano

Differential Revision: D62050671

fbshipit-source-id: 1b6952ed7990d3dd146c958689c289f7e1f5c819