Limit usage of TypeIntersectionConstraint
Summary: Today, type constraints are split across the main type constraint and upperbounds (e.g. retTypeConstraint, returnUBs). The next diff modifies TypeIntersectionConstraint to enable us to merge these together by using a union of a single TypeConstraint and a FixedVector of TypeConstraints. Given that a FixedVector is immutable once initialized, this diff limits the usages of TypeIntersectionConstraint only to places where the constraints are set in stone like func, class prop etc. and excludes it from places where the constraints are being gathered.

Reviewed By: jano

Differential Revision: D62452153

fbshipit-source-id: c92a0420f6b5b6c679ea5e5762a0e6ef40b2e7b9