Append values of unknown types to keysets without punting
Summary:
SetM doesn't guard its top stack input (`DontGuardStack1`), as that input is
a value that is being moved inside an array element or a property. SetM's
handling of appending new element into a keyset tries to use DataTypeSpecific
(which doesn't work) and `PUNT()`s if the key does not have a known DataType.

This causes issues for `HH\Lib\Keyset\keys()`, which obtains keys from an input
array and puts them into a keyset. It knows that these keys are `TInt|TStr`, but
doesn't know the precise DataType. When we inline this function, we always side
exit (except of an empty input array).

Let's handle unknown key types properly and forward them to the C++ helper.

Reviewed By: ricklavoie

Differential Revision: D65459869

fbshipit-source-id: 9889a512fb4e3fa2e0b4c6b859a592a08eea20b0