add `find_one_opt` to WrappedMap
Summary:
WrappedMap is a simple wrapper around the OCaml stdlib's map module. The only way to search through a stdlib map in OCaml using a predicate is by using the `find_first` and `find_first_opt` functions. A limitation of these functions is that they require the predicate to be monotonic: That is, if `a <= b` and `f(a)` holds, `f(b)` must also hold to ensure we find the right values.

This diff introduces a similar function called `find_one_opt` that guarantees the least key matching the predicate being returned, but drops the predicate's monotonicity requirement. Internally, it respects the iteration order of the map's underlying BST and raises an exception on the first element found. Since the iteration of the map's BST passes elements to increasing over, we should also always evaluate to the least key that matches the predicate.

Differential Revision: D65161954

fbshipit-source-id: 7dd495ef081ff2d979283dd15f282cfaa9ad876f