a new way to hugifyText
Summary:
On x86_64 and newer kernels, we use THP on the hot text directly, without manual manipulations.
In places where the above mechanism doesn't work well (e.g., on some aarch64 machines), we have the option of using 2MB hugetlb pages.

The new way of manipulation makes use of mremap(), so that we can prepare the contents of the page before replacing the mapping "atomically". This way, the mappings for the hot text range always contain the right instructions, and we don't have to worry much about things like memcpy() cannot be called during hugifying.

In case mremap() doesn't work for whatever reason, we try to fall back to the old way (which I call restoration in the code). The mechanism worked when I pass invalid addresses in mremap() in an earlier version of this diff.

Reviewed By: ottoni

Differential Revision: D65170134

fbshipit-source-id: 2b22812f8b1f1808cc7a4144e0cd38d2ec3ac072