Add extra-data to InstanceOf[Iface] to signal it can simplify to vtable check
Summary:
In some cases, the InstanceOf op will simplify incorrectly to InstanceOfIfaceVtable, which is invalid when the class in S0 is an interface as we don't set vtables for those https://www.internalfb.com/code/fbsource/[d60a5805dc44b82f5312e72326708fd7f7f96741]/fbcode/hphp/runtime/vm/class.cpp?lines=3972-3974

Specifically, the instance bits check needs to fail and S1's slot needs to not be kInvalidSlot.

This diff adds an extra-data bool to signal that S0's class came from an object that we can then read during simplification.

 ---

A possible different approach would be to refactor the `InstanceOf` IR ops so that S0 is an `Obj` for them (adding new ones that operate on classes) and only simplify those to the vtable check.

Reviewed By: ricklavoie

Differential Revision: D65465985

fbshipit-source-id: 87398d3d9035bf4dd9c726910cc364b7ed0b7f2b