allow different deployments in cli server mode and webserver mode
Summary:
We want to enable setting a different deployment in the cli-server mode than regular
webserver mode in sandbox, namely for the existence requirement of scripts.
This [post](https://fb.workplace.com/groups/890924834786372/permalink/1657938464751668/) details the rationale behind this, with the summary being that in the V0 rollout of
packages/deployments, we would be "hardcoding" the active deployment as config settings
 for cli-server mode and webserver mode in cli.hdf and config.hdf respectively.
This diff adds support for this capability.

Reviewed By: paulbiss

Differential Revision: D64519341

fbshipit-source-id: d91691060140afb85960ddc819deeedba3862ed2