Add option to set name of packages.toml file
Summary:
As packages v2 is rolled out, it would be very helpful to be able to set the name of the packages file to use. This way, we commit a v2 version and can set it in the `HHVM_RELEASE_BUILD_TEST.hdf` file for the HHVM_RELEASE_BUILD_TEST pipeline (D64765686), which our Conveyor's `build-www` node uses, without changing anything in production.

The idea is that option is temporary and will be removed once we're confident with v2 packages and toml file used by it.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D64791301

fbshipit-source-id: 3708b99df3c88d9efb5dd5c85028dcc22e06c44e