switch Set__ to python impl
Summary:
Remove the `_cpp_obj` `std::shared_ptr<std::unordered_set<T>>` in `Set__X`.

This diff does the nasty cython changes to make it all compile!

WARNING: As a result of this diff, thrift-py3 sets that are created in python will be "unordered" instead of key-sorted as is the current behavior for most structs without `cpp.Type{template = "std::unordered_set"}` override.
NOTE: This diff does not change the ordering of `set` that are struct fields. For now, these will retain implicit ordering from the cpp type (`std::set` by default), but this implicit ordering will disappear on migration to thrift-python.

To facilitate testing, `str(my_py3_set)` will output the set in key-sorted order, which may differ from the `__iter__` order for sets that are not created as standalone objects independent of some struct/union.

Reviewed By: yoney

Differential Revision: D65642278

fbshipit-source-id: a110878d9f21b3d6c32ffcef02230e4352704da2