Fix ODR issues around weak symbols for bundled runtime schema
Summary:
This diff removes the default implementation for the weak symbol used for finding runtime schema information of transitively included Thrift files.

If a thrift file is included in more than one file, then we have an ODR violation for the default implementation provided by the weak symbols. In practice, I have witnessed that this causes in some cases, for the default (empty) implementation returning `""` to be picked instead of the strongly linked symbol.

The bug has manifested in weird ways such as different implementations being picked depending on which header files are `#include`-ed, even though they do not cause any changes in the transitively linked-to targets in BUCK.

The fix is to only have forward declarations for the weak symbols with no definitions. This passes ODR since declarations are not definitions, and so we introduce a helper function to deal with the fact that the symbol can now be `nullptr`. Now, build are working again.

Reviewed By: pranavtbhat

Differential Revision: D65635942

fbshipit-source-id: 012e3942c7894c4a35d323a0ffb127dfcf16d23e