Fix dead region block deletion in sortBlocks()
Summary:
While experimenting with the next diff, I accidentally created a region with
some unreachable blocks and saw a crash in sortBlocks(). Turns out while it is
deleting a block and iterates preds or succs, it also removes stuff from them,
corrupting the iterator.

Simplify this logic to remove only the other side of the edge, since we are
nuking the entry in m_data few lines later anyway.

Reviewed By: ricklavoie

Differential Revision: D65801212

fbshipit-source-id: 49846cd0e35397566a26e122d19b0310c9bb3cfe