fix oneway generation for single response, no reply
Summary:
Apparently this flag is only for oneway methods, where if the function returns `void` it's still a reply.

ADDITIONALLY: this fixes the interaction flag, where we were generating a method name for `createXXXInteraction` and we shouldn't, and `starts_interaction = true` should be the case for all methods which return an interaction.

Context: https://www.internalfb.com/diff/D65489886?dst_version_fbid=907093098230055&transaction_fbid=557001307063568

Reviewed By: tlj77

Differential Revision: D65829294

fbshipit-source-id: 4c2dee08cfcff293d64843d8697248f896048f45