Recover state of locals from an immediate dominator
Summary:
If the state of locals was cleared due to unprocessed preds, try to recover it
from the input state of its immediate dominator's successors. Only the state of
locals that were not modified in any path between the immediate dominator and
the current block is recovered.

Preserves state of locals throughout and inside loops, such as the ones in HSL
helpers (e.g. `Vec\map()`). Makes it possible to inline lambdas passed to these
HSL helpers.

This works, because all paths from the entry block to the current block (the
one with unprocessed preds) needs to go through the immediate dominator. The
input state of all idom's successors is used rather than output state of the
idom, since the output state is ambiguous (different block exits might have
different output state).

Since we are processing the current block, the idom block must have been
processed too (otherwise it would not be reachable). Processing of the idom
block have populated the input state of all directly reachable successors
(there must have been at least one for the current block to be reachable).
Unreachable idom successors are safe to ignore, because even if they become
reachable later, the path to them must go through the idom and its another
successor that was directly reachable.

Reviewed By: ricklavoie

Differential Revision: D65849303

fbshipit-source-id: 6a52d24221f8be5b7b235f5e090cb9b9562aeb94