Pass in base service name to ServiceInterceptor
Summary:
Consider:

```
service A {
  void foo();
}

service B extends A {}
```

Today, `ServiceInterceptor` receives service name as "B" and method name "foo". The fact that "A" is where the service is defined is lost.

For ACLs, the defining service is what is important. This passes an additional string to `onRequest` and `onResponse` that includes this information (see fixtures).

Reviewed By: WoodyWangHou

Differential Revision: D65899533

fbshipit-source-id: 2a0e7919d21883d1db029348f2ccce2fe55db31e