Calculate select_queue_duration in HHVM
Summary:
In the future world of web stacking, we need to get all the dyno counters from somewhere else.  This diff:

* `hhvm.select_queue_duration_us.avg.60`: The time a message stays on the queue

* [Original post asking about this](https://fb.workplace.com/groups/1904841443362189/posts/1986341638545502)
* [Discussion about the discrepancy between dyno. and hhvm. counters](https://fb.workplace.com/groups/1904841443362189/posts/2001833150329684)

NOTE: Don't worry about the signal.  This diff is currently on the commit that is running in prod because repo-schema.  When I get a stamp, I'll rebase and ensure everything passes

Reviewed By: binliu19

Differential Revision: D65794349

fbshipit-source-id: e705ad068a8450b47cd1647b51691c6146188f6f