Implement type checking for shape destructure
Summary:
Basic support for type checking shape/tuple destructure. We traverse the shape/tuple tree of the lvalue and generate a fresh type variable for each leaf, and create a type with the same structure as the shape/tuple tree to subtype against the type of the RHS. We ignore some of the complications that are present in the list destructuring code that are only there to deal with holes.

In the future, we might need a constraint to get assignment from dynamic to work.

Reviewed By: andrewjkennedy

Differential Revision: D65891300

fbshipit-source-id: 56f3884050b4169a773a0b0702edf7f1c1d4a1e1