Restrict use of ::class for shape keys at definitions
Summary:
Ban ::class as a shape key on construction and in types.

This functionally shifts `nameof C` closer to string keys than class const keys. To close the loop, and make the switch to string keys, we need to
- ban XHP classname keys (both self and explicit, annoying to model mangling)
- model the root-namespace stripping behavior of the runtime for `\A\C::class = "A\C"`
- cut the type over from TSFclass_const to TSFstring

Then shapes will be able to support multiple `nameofs` and detect collisions.

Reviewed By: viratyosin

Differential Revision: D65865867

fbshipit-source-id: a408fc6274a8a280aac305bdb5798a516ebe7d54