let hphp_context_init initialize transport in xbox, don't null it at end
Summary:
D50249107 cleaned up the dead support for not resetting state across requests in RPC handler.

That persistent context support was the reason we were setting ExecutionContext's transport at the beginning of the handler method and nulling it at the end. Now that A is gone, B is no longer needed either. This will let us access transport in extension request shutdown handlers so I can port Accounting over to an extension.

Reviewed By: binliu19

Differential Revision: D65776568

fbshipit-source-id: dccea9f06394d17a0e00aa31d7137174752746f0