Populate db_version when going straight to MySQL
Summary:
Currently we only populate the `db_version` field in the logging when it is returned as a response attribute for a query.  This means only queries that go through MyRouter get this field included.  However we have this information in the connection when the client connects.  It will be the MyRouter's (or Dvaar's) version when using a proxy, but will be the correct MySQL version when going directly to MySQL (like TAO does).

If we populate this information into the sample first (for connections as well as queries) it can then be overridden for MyRouter/Dvaar when the response attribute comes back.

This change will allow queries going to MySQL to have the `db_version` field set.

In addition we will now have the `db_version` set for successful connections as well.  For direct MySQL connections it will be the MySQL's version.  For proxy connections it will be the MyRouter or Dvaar version - which is less meaningful but theoretically could be helpful.

Differential Revision: D65959868

fbshipit-source-id: 99d6ef76b7c3fbc4bbbdf0f41b14770729198252