Add new fallback for LdCls[Cached]
Summary:
Split of D58392241, trying to narrow down incorrect behavior in is_a for strings and lazy classes.
 ---
The IR instructions for class loading do not support returning null, so we add a new fallback variant `Silent`. We leaned on `loadFail` to fatal or throw instead of proceeding and updating the `ClassCache`, so just add a return after it to give back the nullptr in the `Silent` case.

 Adds new dest kind `DLdCls` that is either `TCls` or `TCls|TNullptr` based on the fallback extra-data

Reviewed By: jano

Differential Revision: D58890532

fbshipit-source-id: aea01d93ed5e3a1874562a640f25f348bd0b9d8e