Add ability to specify empty package
Summary:
Thrift files must have package statements in order to accept file-level annotations. However, adding a package statement currently causes a number of sometimes undesirable effects including specifying URIs for structs (which can't be safely changed in the future) and changing the C++ codegen to use the op library (which can cause build failures).
In order to unblock schema bundling, which requires adding file-level annotations to files without package statements, this diff adds the ability to specify an empty package.
An empty package statement is a no-op in every respect; it is equivalent to not having a package statement at all. This means it can be freely replaced with a nonempty package statement in the future.

Reviewed By: thedavekwon

Differential Revision: D66031837

fbshipit-source-id: 18e74c8766a78f5007da5bf4a9f53eb672fa0682