Add option to output full precision floats in json_encode_with_error
Summary:
HHVM uses 14 digits of precision for floats. Floats that are ieee754 double precision numbers need up to 17 decimal digits to be fully represented without loss of precision. That is already reflected in HHVM `serialize` function that is using 17 decimals for floats (D1202707).

Since JSON is a serialization format, it is good to have an option to serialize floats without loss of precision. Adding new option JSON_FB_FULL_FLOAT_PRECISION to enable that.

Reviewed By: ricklavoie

Differential Revision: D66174523

fbshipit-source-id: a18294b01686bac270c558b71584f9bd954545b5