Filter out built-in lint rules in the presence of take_rules/skip_rules
Summary:
The thrift linter always yielded the built-in ast rules, regardless of the   `--skip_rules` and `--take_rules` flags.

When providing inclusion & exclusion flags, the expectation is to see only those flags. As the built-in rules have the empty string as a name, it's also not possible to filter these by lists above.

In order to opt-in to the old behavior, a `--take_builtin_rules` flag is provided.

Reviewed By: praihan

Differential Revision: D66119471

fbshipit-source-id: 9e5620718d551c4fe7b1ac1e951e3abd64553dd7