Account for man page description of POLLERR
Summary:
This diff refers to our Poll API around sys call `poll`. We've seen a few POLLERR flags in production. The man page says:
```
       POLLERR
              Error condition (only returned in revents; ignored in
              events).  This bit is also set for a file descriptor
              referring to the write end of a pipe when the read end has
              been closed.
```
We account for that by returning `hup = true` if POLLOUT is passed and POLLERR is returned. We also don't consider it an error in that case.

Reviewed By: mheiber

Differential Revision: D66171883

fbshipit-source-id: eb7edad9149862cfb626027ea8cf328119657dc4