Add mechanisms to disable schema const
Summary:
In order to avoid needing weak symbols we would have to ensure the named symbol is always available. This diff handles the case where the path is not well-known, for example because the file is included using a file-relative path instead of a repo-relative path.
The file without a well-known path applies this annotation to its package statement, and is excluded from the list of transitive includes so its constant name does not have to be known.
Since every file using file-relative includes would need this annotation, we instead automatically disable schema const generation for files using such includes as there are a large number and we want to bring that number down but not block schema bundling on that.

Reviewed By: yoney

Differential Revision: D65977818

fbshipit-source-id: 02a1351a2b0fd52332ac068c4f53212b051623e6