Add HH\ReifiedGenerics\get_class_from_type
Summary:
This is meant to offer a replacement for the use of `type_structure(static::class, 'TFoo')['classname']` which otherwise returns a string and requires an implicit string to class conversion to use as a class.
 ---

Adds a new function that's adjacent to HH\ReifiedGenerics\get_classname
https://www.internalfb.com/code/fbsource/[1d9a35d691f6609f58d46f9256ad925c79e97fe0]/fbcode/hphp/runtime/ext/hh/ext_hh.php?lines=436

but is a simple use of the ClassGetTS bytecode.
https://www.internalfb.com/code/fbsource/[387bf2581dbab9a0f9468f1444aec7301226851f]/fbcode/hphp/doc/bytecode.specification?lines=1343
We discard the second resulting vec and leave the class pointer. This is roughly what happens for `new T()` except that we don't support passing reified arguments in `new $c()` expressions.
 ---

The HHBBC types produced by ClassGetTS are too pessimistic, I'd like to address this separately.

https://www.internalfb.com/code/fbsource/[8c6f6a3f38c7aa9a964db6c5649677389512c954]/fbcode/hphp/hhbbc/interp.cpp?lines=2422-2423

I looked at tcprints in repo mode for my test cases, and it does correctly specialize the calls when it has the static type structure vec available.

Differential Revision: D66207622

fbshipit-source-id: 909176fcfd3648413488bcc5097b7ae96b4bd1c0