Add enum<E> type to match enumname<E>
Summary: This adds a type `enum<E>` that expands to `class<BuiltinEnum<E>>` during decling. This mirrors the HHI alias `enumname<E> = classname<BuiltinEnum<E>>`. The type shares the syntax node for class pointer types, just varying the keyword, and its hint representation is just is a variant of the class pointer hint.

Reviewed By: viratyosin

Differential Revision: D66044536

fbshipit-source-id: de9f987e5fe8933907f6449e04b77dff1fee9094