Add targets with abstract types enabled for thrift python library tests
Summary:
# Problem
Currently, tests like `struct_test.py` and `union_test.py` do not enable `experimental_use_abstract_types`. This means there are changes to abstract types that might break these tests and are not covered.

# Solution
To mitigate the risk of breakage to mutable types, create another set of targets that enable abstract types in the tests.

Reviewed By: yoney

Differential Revision: D66275225

fbshipit-source-id: 6cba5796259f549f5e18d812eb1332866eb39910