make skip_n specializable
Summary: Right now in skip and skip_n free functions the recursive calls are to the free function skip. This change is to make the skip calls to the member function of the protocol, to make it easier to customize behavior for a protocol. Now if you call the free function skip_n, skip, or the member function it will all keep going through the member skip function.

Reviewed By: iahs

Differential Revision: D66107283

fbshipit-source-id: 3357d07a6d3254aa956d9e33822829505bec4ad8