Make sure FCall and MemoSet are in the same hhbbc block
Summary:
In memoized wrappers, we currently emit try/catch around the FCall to restore
the implicit context. In case the memoized function produces a constant, hhbbc
reduces the FCall to a constant. But it can't see that constant when processing
the MemoSet during the analysis phase, because we don't run DCE before the
optimize phase that would merge the blocks.

Only in the optimize phase the blocks are merged, the MemoSet propagates the
effect free flag to the MemoGet and the MemoGet gets optimized away with a
constant. But it is too late, as callers won't see the effect free flag and
won't replace the whole call to a memoized function with a constant.

This causes issues with nested memoized functions, where only the body of the
innermost one gets optimized.

By moving MemoSet to the try{} block, this optimization will be performed
during the analysis phase.

It still won't propagate the effect free flag to the return value due to
another issue that will be resolved by the next diff.

Reviewed By: ricklavoie

Differential Revision: D66416069

fbshipit-source-id: d432cee3409069e10b5a90cfdb08c4f8291537fa