Infer that memoized constants are effect free
Summary:
The effect free flag is tracked per the analysis phase and can only go from
true to false. This doesn't work well with memoized constant functions, as the
memoize wrapper uses MemoGet that branches to a block producing the value and
caching it with MemoSet, which is not effect free. There are also implicit
context bytecodes that are not effect free, but their effects is cancelled out
when the function produces a constant.

When MemoSet detects that the preceding FCall was reduced to a constant, it
propagates this information to the MemoGet instruction and its block is
reevaluated. By this time the analysis pass is no longer effect free and
callers of the memoized function cannot optimize the call away. To recover the
effect free flag without harming the monotonicity within the analysis pass, we
need to rerun the analysis pass after the bytecode was updated and the MemoGet
was reduced to a constant. Have a flag that tracks whether this has happened
and rerun the analysis if the flag is set.

Reviewed By: ricklavoie

Differential Revision: D66416178

fbshipit-source-id: d556874a14bd05b168ffc9b752226c425443175e