Handle surprise checks from JIT
Summary:
Do not revert the iterator state and go back to the interpreter to interpret
IterNext and handle the surprise there.

Instead, handle the surprise from JIT, then ReqBindJmp to the iterator body.

Fixes the issue of Xenon misattributing time spent in loops to the interpreter.
Makes it also faster to recover from surprise, as the translation won't need to
be looked up after interpreting IterNext anymore. The cost for this is a
slightly increased code size (call to surprise handler instead of dec).

Reviewed By: ricklavoie

Differential Revision: D66430521

fbshipit-source-id: 7a1972f7617e1d22bbbd143d3ae6899016bbfb1e