Improve check for `case type A = A` to account for type constants
Summary:
This is a follow up to D66248191. The problem with that diff is that it didn't account for type constants / didn't expand type constants. So it would fail to catch that invalid case type:
```
case type A = ?B::T;
class B {
  const type T = A;
}
```

There's currently no code allowing to expand type constants outside of localization. So to account for type constants, we localize the RHS then check for forbidden recursive case types on the localized type.

For this to behave properly, we pass `visibility_behaviour = Always_expand_newtype` which we introduced in the previous diff.

Differential Revision: D66373383

fbshipit-source-id: 30843074cafd9865dfe95ce2b347fb5620f5261c