errors streaming: run error validation with a timeout
Summary: For error streaming, once we've streamed errors, we validate those errors by doing a normal RPC call to get the errors, then compare them. Most of the time this is fast because the typechecker should have finished typechecking. But if in the meantime somehow files change, this can take a while. So we use a timeout for the validation, using Lwt.

Reviewed By: patriciamckenzie

Differential Revision: D66499664

fbshipit-source-id: ffae3d032d347266aa55edbe7f5559e1dcd3798e