refactoring - ClientLsp - ide_service does not need to be a reference - 1/2
Summary: In ClientLsp, the variable `ide_service` is often passed as a reference. Reference are confusing because they seem to indicate that the function taking the ide_service as a reference modifies the ide_service, when it's not the case at all. Removing the reference makes it clear that there are no stateful changes.

Reviewed By: patriciamckenzie

Differential Revision: D66536205

fbshipit-source-id: 51593997329f3ca30208bbcf3f866f26470bdc06