Avoid legacy reflection in testset-based tests
Summary:
This utility is the last use of the deprecated `reflect_module` functionality, and it is only used in two low-value tests that also take up an inordinate amount of time to run and compile.
By hardcoding a random subset of the testset types we can preserve the limited value of these tests (which we may want to separately rethink) while significantly reducing their cost and allowing us to delete the reflection functionality.

Reviewed By: yoney

Differential Revision: D66548840

fbshipit-source-id: 27ab82082ce0763578ce27efa9905692d27d1d42