Remove dead std::to_chars version of float-to-str conversion
Summary:
This removes the dead code path that does float-to-string conversion with `std::to_chars`.

This increased binary size on android because android apps bundle a copy of the C++ standard library in each APK. This was the opposite of the intention to reduce binary size for android by relying on the standard library instead of double-conversion.

Reviewed By: Gownta

Differential Revision: D66378465

fbshipit-source-id: 7ba44c3388e704a867197e366f0e7d2740f0a849