Enforce class<C> type hint shallowly
Summary:
This implements enforcement of the `class<C>` type hint represented as `"class"` in hhas, omitting inner class information for the moment. It closely models the `classname<C>` type hint and admits the same union of string, lazy class, and class, optionally raising a notice for string inputs.

In a type structure, I've reserved kind `T_class_ptr = 33` to represent the type. Since Hack bans `classname<C>` in is/as and reified generics, I'm keeping the same restriction for `class<C>`. The type structure check will throw for is/as, and otherwise admit only lazy class and string. By comparison, `classname<C>` resolves to `T_string`, which admits lazy class and class with an optional notice.

Miscellaneous
- Renamed the Warn and Convert AnnotActions to be a bit clearer
- Renamed `kUnionTypeClass` bit to `kUnionTypeSubObject` to match the AnnotType that sets it

Reviewed By: jano

Differential Revision: D58316371

fbshipit-source-id: d47488e861bc9081dc23ba0a23748364d9075d7f