Add class_or_classname runtime type
Summary: Add a type similar to the `vec_or_dict` type that admis strings, lazy classes, and classes without any of the logging flags. This type will be banned from is/as and reified generics in the typechecker, matching the behavior of classname<T>. Its main purpose is to support cases that conflate the types, conversion functions, and to enable breaking the `class<T> <: classname<T>` subtype relation.

Reviewed By: jano

Differential Revision: D65913769

fbshipit-source-id: d66698bb33e3a29c0e2a0dc2e0be2e5657f1a404