refactoring - ClientLsp - ide_service does not need to be a reference - 2/2
Summary: In ClientLsp, the variable `ide_service` is often passed as a reference. Reference are confusing because they seem to indicate that the function taking the ide_service as a reference modifies the ide_service, when it's not the case at all. Removing the reference makes it clear that there are no stateful changes.

Reviewed By: patriciamckenzie

Differential Revision: D66536204

fbshipit-source-id: 5178f724f3a527c9246d20d469c3b239e01d3b07