Add flag to record instantiation
Summary:
Currently we instantiate `id<>` with fresh type variables so that we get a type error on the second application. Instead of this, we'd like to type function pointers which are explicitly unapplied as polymorphic function types which are handled during subtyping. With this approach we are free to apply `$f` at any argument repeatedly:

```
function id<T>(T $x): T  { return $x }

function use_id(): void {
  $f = id<>;
  $_ = $f('a');
  $_ = $f(1);
}
```

We don't yet have a way of representing universally quantified type so we need some other way to ensure we can determine whether a function type is instantiated. This diff adds a flag for this purpose.

Reviewed By: andrewjkennedy

Differential Revision: D66543480

fbshipit-source-id: 69c2dc4a58645e1c20e43a2933f2d948ea69107d