Refactor AST of inout arguments
Summary:
Change the representation of arguments to Call nodes from `param_kind * expr` to `type Argument = Anormal of expr | Ainout of pos * expr`. This is isomorphic, so this diff shouldn't change any behaviour.  The motivation is to set of a subsequent diff that restricts the type of `Ainout` to those lvalues that can actually show up.

It was prevalent to match on `(_, Expr ...)` and not care if the argument was inout or not. Most of those have become explicit calls to `Aast_utils.arg_to_expr`, although sometimes the pattern matching structure had to be modified when the arg itself was nested inside of another pattern.

Reviewed By: andrewjkennedy

Differential Revision: D66544188

fbshipit-source-id: d17f4f71952f99cf4db43fcfa6fe65657f52134c