Recursive case types: more complete subtyping
Summary:
Now that we've checked for contractiviness in the previous diff, we can implement the subtyping algo for recursive types from Pierce (fig. 21-4 p. 305 of TAPL), which essentially says that when recursing in subtyping, that means the subtyping relationship holds.

When case types are not contractive, e.g. `case type A = A`, then subtyping relationships like `string <: A` will now be accepted. I believe this is fine so long we reject the wellformedness of A.

Reviewed By: andrewjkennedy

Differential Revision: D66451606

fbshipit-source-id: a9495d53dc0b2337d9dcd42396c9c10a397c9e2a