allow customizing Struct/Union source module name
Summary:
In order to make thrift-py3 have a `python_library` `types.py` file on trunk, need to be able to "fake" the module name.

For instance, if we want to put it in `types_impl.pyx`, then re-export them through `.types`, the generated `Struct` and `Union` need to pretend like they're from `.types` module to avoid breaking customer code that relies on this.

This diff itself is a no-op in terms of the class's apparent `__module__`. It just serves to validate we can customize the name, de-risking next diffs.

Reviewed By: createdbysk

Differential Revision: D66516699

fbshipit-source-id: 5e3450acd00a65019b53602afd3f81f6c10d016b