Struct-less Patch
Summary:
This diff implemented `protocol::tagged_patch<Tag>`. It provides the same API as static patch, but it does not require the generation of Patch struct.

Internally, we implemented

1. `TaggedPatchRef<Tag>`: A wrapper that refers to the underlying dynamic patch without owning it. It uses Tag to provide type safe APIs.
2. `Owned<TaggedPatchRef<Tag>>`: This contains both `TaggedPatchRef` and the underlying DynamicPatch, so that it owns the data while providing the same APIs.

The reason we need `TaggedPatchRef<Tag>` is to support APIs like `patchByKey(...)` and `patchIfSet(...)` which returns another DynamicPatch, thus we need to create a class that refers this patch.

Differential Revision: D66561994

fbshipit-source-id: 86340211fe32fd7abf16be9aa2db7f8a8c477f39