3/[thrift] Migrate from THeader::TRANSFORMS to TTransform in THeader.
Summary:
This diff replaces all transform usage of `uint16_t` with `TTransform`. APIs using the `uint16_t` transforms are marked deprecated.

Meta:

Note that this diff was landed (D65832231) and reverted (D66329820). The root cause was that another conflicting diff (D65494301) was landed at the same time. It was also reverted (D66327051).

The two reverted diffs were individually safe, but unsafe when combined. We are re-landing the other diff first, and this diff was modified to take those changes into account.

Reviewed By: felixhandte

Differential Revision: D66372677

fbshipit-source-id: 6077105e3f88f46b35967564531ee3374530c9ed