Lift assignment out of Binop expressions
Summary:
Move assignments to their own expression constructor, out of Binop. This will allow us to change the type of the lvalue from `expr` to a custom lvalue type that is not comparable with expressions, but will have some overlap. In particular, we want to include more patterns than just lists, tuples and shapes that  also happen to be expressions.

Also, once we remove assignments in expressions from the hhvm test suite, we can move assignment from expr to stmt.

Reviewed By: madgen

Differential Revision: D66582409

fbshipit-source-id: 2ee5e769c12bd4e95b0f696a9429ef6ee3fecb0f