Error streaming: log more in case of payload size assertion failure
Summary:
This is a follow up of D66251374.
Context: When deserializing a chunck of the errors file, we first read a preamble that contains the size of the chunk. To avoid allocating too much memory in case of a serialization error, we assert that the size is below a certain limit. That limit is currently as 20 million bytes. Based on [this comment](https://www.internalfb.com/code/fbsource/[7fc0b94002f166e57fcf0887491637c40735428a]/fbcode/hphp/hack/src/client_and_server/server_progress.ml?lines=315), I suspect this is an error and Lucian intended to set it at 20GB, but I can't be sure.

Now that we've started to roll out, we see assertion failures: https://fburl.com/scuba/hh_server_events/o213m4ve.

This diff does two things:
* it raises the limit to 200 GB
* it logs more about the payload in case the limit is exceeded.

Reviewed By: andrewjkennedy

Differential Revision: D66822865

fbshipit-source-id: 3d921cee69893f678000ff00750fca2498ddb3a4